Added Transport and Performance capabilities
Updated the refigur8tor iPhone5 template. It now
included 2 configured multiballs updated to used the codec.
The transport controls on top can be used to trigger Reason.
The buttons are not synced and therefore don't necessarily
match the state in Reason, but work as expected.

The acecelerometer is experimental for the Expression Performance control.